Move thank-you functions to composite-checkout dir
